Move unittest utility interfaces to Utils.h
